+++ src/ArduinoJson/Collection/CollectionImpl.hpp
@@ -160,4 +160,20 @@ inline size_t CollectionData::size() const {
   return slotSize(_head);
 }
 
+template <typename T>
+inline void movePointer(T*& p, ptrdiff_t offset) {
+  if (!p) return;
+  p = reinterpret_cast<T*>(
+      reinterpret_cast<void*>(reinterpret_cast<char*>(p) + offset));
+  ARDUINOJSON_ASSERT(isAligned(p));
+}
+
+inline void CollectionData::movePointers(ptrdiff_t stringDistance,
+                                         ptrdiff_t variantDistance) {
+  movePointer(_head, variantDistance);
+  movePointer(_tail, variantDistance);
+  for (VariantSlot* slot = _head; slot; slot = slot->next())
+    slot->movePointers(stringDistance, variantDistance);
+}
+
 }  // namespace ARDUINOJSON_NAMESPACE