+++ src/ArduinoJson/Document/DynamicJsonDocument.hpp
@@ -11,12 +11,16 @@
 namespace ARDUINOJSON_NAMESPACE {
 
 struct DefaultAllocator {
-  void* allocate(size_t n) {
-    return malloc(n);
+  void* allocate(size_t size) {
+    return malloc(size);
   }
 
-  void deallocate(void* p) {
-    free(p);
+  void deallocate(void* ptr) {
+    free(ptr);
+  }
+
+  void* reallocate(void* ptr, size_t new_size) {
+    return realloc(ptr, new_size);
   }
 };